requirements.txt
playwight
pytest
allure-pytest
ytest-asyncio